Changes by nr1sl

+++ src/Voertuig.cpp
@@ -8,7 +8,7 @@ const string &Voertuig::getBaan() {
     return baan;
 }
 
-int Voertuig::getPositie() {
+double Voertuig::getPositie() {
     return positie;
 }
 
@@ -16,7 +16,7 @@ void Voertuig::setBaan(const string &name) {
     baan = name;
 }
 
-void Voertuig::setPositie(int number) {
+void Voertuig::setPositie(double number) {
     positie = number;
 }
